Wish Saya Backend 90%, Transaksi Saya Backend 90%, Backend Detail Transaksi 50%, fix" bbrp hal

+++ app/Http/Controllers/PaymentController.php
@@ -29,8 +29,8 @@ class PaymentController extends Controller
             foreach($payment_items as $pi){
                 // $total_oti = $request->total_oti;
                 // $total_itu = $request->total_itu;
-                $total_oti = 40000;
-                $total_itu = 25000;
+                $total_oti = 40000; // biaya ongkir origin to indonesia
+                $total_itu = 25000; // biaya ongkir indonesia to user
                 // $payment_item->total_oti = $total_oti;
                 // $payment_item->total_itu = $total_itu;
                 $total_payment = $pi->total_price + $total_oti + $total_itu + $pi->total_fee;
@@ -92,7 +92,9 @@ class PaymentController extends Controller
                     ]);
                 }
 
-                Cart_Item::where('id', $payment_item->cart_item_id)->delete();
+                if($payment_item->cart_item != NULL){
+                    Cart_Item::where('id', $payment_item->cart_item_id)->delete();
+                }
 
                 $payment = Payment::where('user_id', $user->id)->update([
                     'total_payment' => $request->grand_total,
